requirements.txt
Django>=3.1,<4.0
django-cors-headers>=3.5.0
djangorestframework>=3.12.2
drf-yasg
selenium
gunicorn
django-heroku